Replace `ParseError` with generic `SyntaxError`
This commit removes `ParseError` and changes the `Parser` so that it
operates directly over a `TokenStream`.

Consequently, this means that the parser will encounter fallible next
token results and thus propagate `SyntaxError`s.

With these changes it is no longer necessary to collect and separate
tokens from errors before parsing them into an expression AST.