feat: support return statements
Note that there used to be a bug with the way that the interpreter
context has been set up which caused the `return_stmt/test_case_1` to
fail.

The problem was that the initial context had as its environment an
already scoped global environment. The test case exercised a recursive
function declaration lookup and because we don't support environment
closures yet, it could not find a function in globals (because in was
defined in the child environment of globals).

Temporarily this has been fixed by using globals as the initial
environment, but will be properly handled by closures in future commits.

Finally, note that due to this change, and due to preparation for
closures, environments are now heap-allocated and ref-counted.