refactor: make lexer available for the VM
This commit just moves some modules around. Specifically it makes the
`Lexer` and `(Lex)Token` structures easily accessible in the VM-based
interpreter.

The errors have been moved as well so in the future `RuntimeError` could
be unified. This will, however, require the bytecode `Chunk` to be
slightly generalized to carry not just `Line`s, but rather full `Span`s.